fix: rename styles, fix callback
