+++ src/components/BoardComponents/styles/index.ts
