+++ src/pages/Board/index.tsx
@@ -4,17 +4,8 @@ import { useParams } from 'react-router-dom';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { getBoardById } from '../../requests';
 import { removeAllColumns } from '../../store/boardSlice';
-
-const BoardTitle = React.lazy(() =>
-  import('../../components/BoardComponents/BoardTitle').then((module) => ({
-    default: module.BoardTitle,
-  }))
-);
-const ColumnList = React.lazy(() =>
-  import('../../components/BoardComponents/ColumnsList').then((module) => ({
-    default: module.ColumnList,
-  }))
-);
+import { BoardTitle } from '../../components/BoardComponents/BoardTitle';
+import { ColumnList } from '../../components/BoardComponents/ColumnsList';
 
 export const Board = () => {
   const { id } = useParams();
@@ -22,7 +13,7 @@ export const Board = () => {
   const {
     authUser: { token },
   } = useAppSelector((state) => state.authUser);
-  const { error, columns, boardTitle, boardDescription } = useAppSelector(
+  const { error, columns, boardTitle, boardDescription, isLoading } = useAppSelector(
     (state) => state.boardState
   );
 
@@ -33,14 +24,13 @@ export const Board = () => {
     };
   }, [dispatch, token, id]);
 
+  if (isLoading) return <CircularProgress />;
   if (error) return <h3>{error}</h3>;
 
   return (
     <>
-      <React.Suspense fallback={<CircularProgress />}>
-        <BoardTitle title={boardTitle} description={boardDescription} token={token} id={id} />
-        <ColumnList columns={columns} token={token} boardId={id} />
-      </React.Suspense>
+      <BoardTitle title={boardTitle} description={boardDescription} token={token} id={id} />
+      <ColumnList columns={columns} token={token} boardId={id} />
     </>
   );
 };
